Fixes team invite accept and team invite decline test [Lunn]
TeamInviteTest set merged into TeamDetailActivityTest set
